fix warning

+++ lap_solver/core/cuda/lap_worksharing.h
@@ -20,7 +20,7 @@ namespace lap
 				cudaDeviceProp deviceProp;
 				cudaGetDeviceCount(&device_count);
 
-				for (int current_device = 0; ((current_device < device_count) && (device.size() < max_devices)); current_device++)
+				for (int current_device = 0; ((current_device < device_count) && ((int)device.size() < max_devices)); current_device++)
 				{
 					cudaGetDeviceProperties(&deviceProp, current_device);
